dont touch the red lines

+++ dotNet5782_0286_2346/BL/DroneInCharging.cs
@@ -12,7 +12,7 @@ namespace IBL
         public class DroneInCharging
         {
             public int Id { get; set; }
-            public int BatteryStatus { get; set; }
+            public double BatteryStatus { get; set; }
             public override string ToString()
             {
                 return this.ToStringProperty();
